+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- Change `xDomainMin` / `xDomainMax` & y equivalents to `xMin` / `xMax` etc. ([#95](https://github.com/graph-paper-org/graph-paper/pull/95))
 - Refactor width / height of `DataGraphic` component to reactively update ([#83](https://github.com/graph-paper-org/graph-paper/pull/82))
 
 ## 0.0.0-alpha.8